requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
astor==0.8.1
attrs==23.2.0
autorepr==0.3.0
bandit==1.7.9
black==24.4.2
cattrs==23.2.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
ConfigUpdater==3.2
coverage==7.6.0
darglint==1.8.1
dictdiffer==0.9.0
docutils==0.21.2
dpath==2.2.0
eradicate==2.3.0
flake8==7.1.0
flake8-annotations-complexity==0.0.8
flake8-bandit==4.1.1
flake8-broken-line==1.0.0
flake8-bugbear==24.4.26
flake8-builtins==2.5.0
flake8-commas==2.1.0
flake8-comprehensions==3.15.0
flake8-debugger==4.1.2
flake8-docstrings==1.7.0
flake8-eradicate==1.5.0
flake8-isort==6.1.1
flake8-mypy==17.8.0
flake8-quotes==3.4.0
flake8-rst-docstrings==0.3.0
flake8-string-format==0.3.0
flatten-dict==0.4.2
frozenlist==1.4.1
furl==2.1.3
gitdb==4.0.11
gitignore_parser==0.1.11
GitPython==3.1.43
identify==2.6.0
idna==3.7
iniconfig==2.0.0
isort==5.13.2
jmespath==1.0.1
loguru==0.7.2
markdown-it-py==3.0.0
marshmallow==3.21.3
marshmallow-polyfield==5.11
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.3.0
multidict==6.0.5
mypy==1.11.0
mypy-extensions==1.0.0
nitpick==0.35.0
orderedmultidict==1.0.1
packaging==24.1
pathspec==0.12.1
pbr==6.0.0
pep8-naming==0.13.3
platformdirs==4.2.2
pluggy==1.5.0
pycodestyle==2.12.0
pydocstyle==6.3.0
pyflakes==3.2.0
Pygments==2.18.0
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
python-slugify==8.0.4
PyYAML==6.0.1
requests==2.32.3
requests-cache==1.2.1
restructuredtext_lint==1.4.0
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
setuptools==72.1.0
six==1.16.0
smmap==5.0.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
stevedore==5.2.0
StrEnum==0.4.15
text-unidecode==1.3
toml==0.10.2
tomlkit==0.13.0
typing_extensions==4.12.2
url-normalize==1.4.3
urllib3==2.2.2
wemake-python-styleguide==0.19.2
yarl==1.9.4
